Renamed the class variables as follows: 1)word to targetWord 2)g to guessesMade 3)wrong to guessesRemaining 4)h to hintsRemaining 5)got to indexOfLettersGuessedCorrectly 6)not to indexOfLettersNotGuessedCorrectly
The g and h constructor arguments were renamed to "maxGuesses" and "maxHints" to make the purpose of these variables clear.

All the class attributes were made private and getter methods were created for each attribute.
Scanner sc was removed as a class attribute and moved inside the guessLetter function. It does not make any sense to make "sc" a class variable as it is used only to read input data and does not tell us about the property of the class. The comments "The game state" was removed as it is a redundant comment. The purpose of the class is evident from its name "GameState". The comment "letters" was also removed because it does not make any sense. The comments "guesses made", "guesses remaining" were deleted because the purpose of the variables is clear from their name "guessesMade" and "guessesRemaining". The useless commented out code "System.out.println(missing)" was also removed as it serves no purpose.